Week #2 - Day #4 - Python and Database/Assignments/Exercises XP/requirements.txt
psycopg[binary]==3.2.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4